+++ tests/testthat/test-select-helpers.R
@@ -1,8 +1,12 @@
 context("select-helpers")
 
+test_that("no set variables throws error", {
+  expect_error(starts_with("z"), "Variable context not set")
+})
+
 test_that("failed match removes all columns", {
-  set_current_vars(c("x", "y"))
-  on.exit(reset_current_vars())
+  old <- set_current_vars(c("x", "y"))
+  on.exit(set_current_vars(old))
 
   expect_equal(starts_with("z"), integer(0))
   expect_equal(ends_with("z"), integer(0))
@@ -13,8 +17,8 @@ test_that("failed match removes all columns", {
 
 
 test_that("matches return integer positions", {
-  set_current_vars(c("abc", "acd", "bbc", "bbd", "eee"))
-  on.exit(reset_current_vars())
+  old <- set_current_vars(c("abc", "acd", "bbc", "bbd", "eee"))
+  on.exit(set_current_vars(old))
 
   expect_equal(starts_with("a"), c(1L, 2L))
   expect_equal(ends_with("d"),   c(2L, 4L))